moved backgroud sub to helpers file
